refactor(kernel): inline local git process adapter
